Mount node_modules folder locally on the guest vm
This would prevent an interference with native modules that were compiled on host (of presumably different architecture) and won't work on guest vm anyway.

It also fixes the problem with npm and windows hosts by elliminating the need in symlinks and working around Maximum Path Length (since modules will never hit the host).