Unbuffer Python output for live ansible output
Before that patch, whole stdout of `ansible-playbook` was returned after playbook was completely finished. By unbuffering python output we allow it to appear live in the terminal.

This helps tremendously from the UX standpoint as before that you had to wait for play to finish before you would know the results.